refactor: mockdata.ts 삭제 및 새로 정의
